rgw multisite: enable more than 1 realm per cluster
Make it so that more than one realm, zonegroup,
or zone can be created during a run of the rgw
multisite ansible playbooks.

If the rgw multisite variables are declared in
group_vars then only 1 zone, zonegroup, and realm
will be created.

If variables are changed in hostvars for each rgw
host in the inventory then realms, zones, and
zonegroups are created on a more finegrained basis.

Zones and zonegroups can be created as either
default, master, or neither.

Signed-off-by: Ali Maredia <amaredia@redhat.com>